Move `ChatDeletionDialog` to its own file.
